created custom ambassador response status  exception and added exceptions to ambassador controller methods

+++ src/main/java/com/zipcode/controllers/AmbassadorController.java
@@ -1,90 +1,80 @@
 package com.zipcode.controllers;
 
 
-import com.zipcode.exceptions.ResourceNotFoundException;
+import com.zipcode.exceptions.AmbassadorNotFoundException;
 import com.zipcode.models.Ambassador;
 import com.zipcode.services.AmbassadorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.ResourceAccessException;
+
+import java.util.Optional;
 
 @RestController
 public class AmbassadorController {
 
-    @Autowired
+
     private AmbassadorService ambassadorService;
+
+    @Autowired
+    public AmbassadorController(AmbassadorService ambassadorService) {
+        this.ambassadorService = ambassadorService;
+    }
 
     @PostMapping("/ambassador/create")
     public ResponseEntity<Boolean> newAmbassador(@RequestBody Ambassador ambassador) {
-        try {
-            ambassadorService.newAmbassador(ambassador);
-        } catch (ResourceNotFoundException e) {
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        ambassadorService.newAmbassador(ambassador);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
-        return new ResponseEntity<>(HttpStatus.OK);
-}
 
     @DeleteMapping("/ambassador/delete/{id}")
     public ResponseEntity<Boolean> deleteAmbassador(@PathVariable Long id) {
-        try {
-            ambassadorService.deleteAmbassador(id);
-        } catch (ResourceNotFoundException e)   {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-
+        if (!ambassadorService.ambassadorExists(id)) {
+            throw new AmbassadorNotFoundException();
         }
+        ambassadorService.deleteAmbassador(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping("/ambassador/update/{id}")
     public ResponseEntity<Boolean> updateAmbassador(@PathVariable Long id, @RequestBody Ambassador ambassador) {
-        try {
-            ambassadorService.updateAmbassador(id, ambassador);
-        }   catch (ResourceNotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (!ambassadorService.ambassadorExists(id)) {
+            throw new AmbassadorNotFoundException();
         }
+        ambassadorService.updateAmbassador(id, ambassador);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/ambassador/{id}")
     public ResponseEntity<Ambassador> findById(@PathVariable Long id) {
-        try {
-            ambassadorService.findById(id);
-        }   catch (ResourceNotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (!ambassadorService.ambassadorExists(id)) {
+            throw new AmbassadorNotFoundException();
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(ambassadorService.findById(id), HttpStatus.OK);
     }
 
     @GetMapping("/ambassadors")
     public ResponseEntity<Iterable<Ambassador>> findAll() {
-        try {
-            ambassadorService.findAll();
-        }   catch (ResourceNotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(ambassadorService.findAllAmbassadors(), HttpStatus.OK);
     }
 
     @GetMapping("/ambassadors/name/{firstName}")
-    public ResponseEntity<Ambassador> findByFirstName(@PathVariable String firstName) {
-        try {
-            ambassadorService.findByFirstName(firstName);
-        } catch (ResourceNotFoundException e)   {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<Iterable<Ambassador>> findAllByFirstName(@PathVariable String firstName) {
+        Iterable<Ambassador> ambassadors = ambassadorService.findAllAmbassadorsByFirstName(firstName);
+        if (ambassadors == null) {
+            throw new AmbassadorNotFoundException();
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(ambassadors, HttpStatus.OK);
     }
 
     @GetMapping("ambassadors/last/{lastName}")
-    public ResponseEntity<Ambassador> findByLastName(@PathVariable String lastName) {
-        try {
-            ambassadorService.findByLastName(lastName);
-        }   catch (ResourceNotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<Iterable<Ambassador>> findAllByLastName(@PathVariable String lastName) {
+        Iterable<Ambassador> ambassadors = ambassadorService.findAllAmbassadorsByLastName(lastName);
+        if (ambassadors == null) {
+            throw new AmbassadorNotFoundException();
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(ambassadors, HttpStatus.OK);
     }
